Minimize diffs by keeping all in one file
